cleanup
git-svn-id: https://subversion.assembla.com/svn/kinvey/ios-library/trunk@5098 868178e2-d38b-4ba8-a603-825f91c87007

+++ KinveyKit/KinveyKit/Source/KCSResourceStore.m
@@ -61,7 +61,7 @@
 {
     NSArray *objectsToProcess = [NSArray wrapIfNotArray:object];
     
-    int totalObjects = objectsToProcess.count;
+    NSUInteger totalObjects = objectsToProcess.count;
     if (totalObjects == 0) {
         completionBlock(nil, nil);
     }
@@ -150,7 +150,7 @@
 - (void)queryWithQuery:(id)query withCompletionBlock: (KCSCompletionBlock)completionBlock withProgressBlock: (KCSProgressBlock)progressBlock
 {
     NSArray *objectsToProcess = [NSArray wrapIfNotArray:query];
-    int totalObjects = objectsToProcess.count;
+    NSUInteger totalObjects = objectsToProcess.count;
     if (totalObjects == 0) {
         completionBlock(nil, nil);
     }
@@ -203,7 +203,7 @@
 - (void)removeObject:(id)object withCompletionBlock: (KCSCompletionBlock)completionBlock withProgressBlock: (KCSProgressBlock)progressBlock
 {
     NSArray *objectsToProcess = [NSArray wrapIfNotArray:object];
-    int totalObjects = objectsToProcess.count;
+    NSUInteger totalObjects = objectsToProcess.count;
     if (totalObjects == 0) {
         completionBlock(nil, nil);
     }
